+++ extensions/itranslate/CHANGELOG.md
@@ -1,5 +1,11 @@
 # iTranslate Changelog
 
+## [Add New Translation Service Providers & New icon] - 2022-08-26
+
+- Add Google Could Translation option
+- Add Aliyun option
+- New icon
+
 ## [Optimize Experience] - 2022-08-18
 
 - Replace List.throttle with setTimeout to optimize the user input experience